🔥 pref : API 기능별로 분리 및 경로 수정
